requirements.txt
folium==0.19.6
matplotlib==3.10.3
pandas==2.3.0
python-dotenv==1.1.0
Requests==2.32.3
schedule==1.2.2
streamlit==1.45.1
streamlit_folium==0.25.0
